requirements.txt
Django==5.2
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-filter==24.3
django-storages==1.14.4
psycopg2-binary==2.9.9
Pillow==10.4.0
redis==5.0.8
celery==5.4.0
django-celery-beat==2.7.0
channels==4.1.0
channels-redis==4.2.0
PyJWT==2.9.0
google-auth==2.34.0
google-auth-oauthlib==1.2.1
stripe==10.8.0
pandas==2.2.2
numpy==2.1.1
plotly==5.23.0
python-decouple==3.8
boto3==1.35.16
gunicorn==23.0.0
whitenoise==6.7.0
dj-database-url==2.2.0